+++ core.h
@@ -1,14 +1,14 @@
 #ifndef CORE_H
 #define CORE_H
 
-void init_pos(double* rx, double* ry, double* rz, const double rho);
+void init_pos(float* rx, float* ry, float* rz, const float rho);
 
-void init_vel(double* vx, double* vy, double* vz, double* temp, double* ekin);
+void init_vel(float* vx, float* vy, float* vz, float* temp, float* ekin);
 
-void velocity_verlet(double* rx, double* ry, double* rz, double* vx,
-                     double* vy, double* vz,
-                     double* fx, double* fy, double* fz, double* epot,
-                     double* ekin, double* pres, double* temp, const double rho,
-                     const double V, const double L);
+void velocity_verlet(float* rx, float* ry, float* rz, float* vx,
+                     float* vy, float* vz,
+                     float* fx, float* fy, float* fz, float* epot,
+                     float* ekin, float* pres, float* temp, const float rho,
+                     const float V, const float L);
 
 #endif
